requirements.txt
feedparser>=6.0
requests>=2.31
anthropic>=0.40
python-dotenv>=1.0
pyyaml>=6.0
rich>=13.0
schedule>=1.2
flask>=3.0
slack-sdk>=3.0
gunicorn>=21.0
